\name{ipsecr-defunct}
\alias{ipsecr-defunct}
\alias{proxyfn0}

\title{Defunct Functions in Package \pkg{ipsecr}}

\description{

These functions are no longer available in \pkg{ipsecr}.

}

\usage{

# Defunct in 1.2.0 (2022-08)

proxyfn0()

}

\details{

proxyfn0 was removed without warning in ipsecr 1.2.0. 
Use \code{\link{proxyfn1}} or \code{\link{proxy.ms}}. 

}

\seealso{

\code{\link{ipsecr-deprecated}}

}
