\name{ipsecr-deprecated}
\alias{ipsecr-deprecated}

\title{Deprecated Functions in Package \pkg{ipsecr}}

\description{

These functions will be removed from future versions of \pkg{ipsecr}.

}

\details{

No functions are deprecated at this point.
}

\seealso{

\code{\link{ipsecr-defunct}}, 

}

